# Models/Property.cs
namespace RealEstateListingAPI.Models
{
    public class Property
    {
        public int Id { get; set; }
        public string Title { get; set; }
        public string Description { get; set; }
        public decimal Price { get; set; }
        
        public byte[]? Image { get; set; }
    }
}